\name{TripleR-package}
\alias{TripleR-package}
\alias{TripleR}
\docType{package}
\title{Triple R: A Package for Round-Robin Analyses Using R}
\description{
The function RR performs Social Relation Model (SRM) analyses based on a single round-robin group. These analyses are either based on one manifest variable, one latent construct measured by two manifest variables, two manifest variables and their bivariate relations, or two latent constructs each measured by two manifest variables.
}
\details{
\tabular{ll}{
Package: \tab TripleR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-04-27\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

For further information of usage see \code{\link{RR}}
}
\author{Stefan C. Schmukle, Felix D. Schoenbrodt & Mitja D. Back
}
\references{
Back, M. D., Schmukle, S. C., & Egloff, B. (2009). From first sight to friendship: A longitudinal social relations analysis of stability and change in interpersonal attraction. Manuscript submitted for publication. 

Kenny, D. A. (1994). Interpersonal perception: A social relations analysis. New York: Guilford Press.

Lashley, B. R., & Bond, C. F., Jr. (1997). Significance testing for round robin data. Psychological Methods, 2, 278-291.
}

\keyword{package}