\name{TripleR-package}
\alias{TripleR-package}
\alias{TripleR}
\docType{package}
\title{Triple R: A Package for Round-Robin Analyses Using R}
\description{
Social Relation Model (SRM) analyses for single or multiple round-robin groups are performed. These analyses are either based on one manifest variable, one latent construct measured by two manifest variables, two manifest variables and their bivariate relations, or bivariate relations between two latent constructs each measured by two manifest variables. Within-group t-tests for variance components are provided for single group, between-groups t-tests for multiple groups. Preliminary handling for missing values is provided.
}

\details{
For further information of usage see \code{\link{RR}} or the built in pdf document `How to use TripleR'. You can open this document either by on this link: \url{../doc/TripleR.pdf}, or you can browse all included vignettes by opening the index of the package documentation (scroll down to the very end of this page and click on `Index'; than click on `Overview of user guides and package vignettes').

If you use this package in your research, please cite it as:
Schmukle, S. C., Schönbrodt, F.D., & Back, M. D. (2010). TripleR: A package for round robin analyses using R (version 1.0). Retrieved from http://www.persoc.net/ToolBox/TripleR.
}

\author{Stefan C. Schmukle, Felix D. Schönbrodt & Mitja D. Back
}
\references{
Back, M. D., Schmukle, S. C., & Egloff, B. (in press). From first sight to friendship: A longitudinal social relations analysis of stability and change in interpersonal attraction. \emph{European Journal of Personality}. 

Bonito, J. A., & Kenny, D. A. (2010). The measurement of reliability of social relations components from round-robin designs. Personal Relationships, 17, 235 - 251.

Kenny, D. A. (1994). Interpersonal perception: A social relations analysis. New York: Guilford Press.

Lashley, B. R., & Bond, C. F., Jr. (1997). Significance testing for round robin data. Psychological Methods, 2, 278-291.

Schönbrodt, F. D., Back, M. D., & Schmukle, S. C. (in prep). \emph{TripleR: An R Package for Advanced Social Relations Analyses.} Manuscript in preparation.
}

\keyword{package}